pyproject.toml
[project]
name = "active-fed"
version = "0.1.0"
description = "Active Federated Learning on CartPole with PPO, Kubeflow, and MLflow"
requires-python = ">=3.10"
dependencies = [
    "torch==2.2.2",
    "gymnasium>=0.29.0",
    "numpy>=1.26.0,<2.0.0",
    "scipy>=1.12.0",
    "mlflow>=2.11.0,<3.0.0",
    "pyarrow<15.0.0",
    "minio>=7.2.0",
    "kfp>=2.7.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "boto3>=1.42.59",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Route torch through PyTorch's own CPU wheel index
# which ships macosx_10_15_x86_64 wheels (unlike PyPI releases that need macosx_11_0+)
[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
